Make runWaiSession part of the public interface
While using state machine testing from hedgehog, it's useful to build up a
session in a `PropertyT WaiSession` and then run the session at the end. In
order to do that access to `runWaiSession` is needed. It can currently be
accessed by importing the `Internal` module, but I think making it more
prominent is useful.